+++ hyp3lib/fetch.py
@@ -1,8 +1,10 @@
 """Utilities for fetching things from external endpoints"""
-
+import cgi
 import logging
+from os.path import basename
 from pathlib import Path
-from typing import Union
+from typing import Optional, Tuple, Union
+from urllib.parse import urlparse
 
 import requests
 from requests.adapters import HTTPAdapter
@@ -22,7 +24,20 @@ def write_credentials_to_netrc_file(username: str, password: str,
             f.write(f'machine {domain} login {username} password {password}\n')
 
 
-def download_file(url: str, directory: Union[Path, str] = '.', chunk_size=None, retries=2, backoff_factor=1) -> str:
+def _get_download_path(url: str, content_disposition: str = None, directory: Union[Path, str] = '.'):
+    filename = None
+    if content_disposition is not None:
+        _, params = cgi.parse_header(content_disposition)
+        filename = params.get('filename')
+    if not filename:
+        filename = basename(urlparse(url).path)
+    if not filename:
+        raise ValueError(f'could not determine download path for: {url}')
+    return Path(directory) / filename
+
+
+def download_file(url: str, directory: Union[Path, str] = '.', chunk_size=None, retries=2, backoff_factor=1,
+                  auth: Optional[Tuple[str, str]] = None) -> str:
     """Download a file
 
     Args:
@@ -31,28 +46,26 @@ def download_file(url: str, directory: Union[Path, str] = '.', chunk_size=None, 
         chunk_size: Size to chunk the download into
         retries: Number of retries to attempt
         backoff_factor: Factor for calculating time between retries
+        auth: Username and password for HTTP Basic Auth
 
     Returns:
         download_path: The path to the downloaded file
     """
     logging.info(f'Downloading {url}')
 
-    try:
-        download_path = Path(directory) / url.split("/")[-1]
-    except AttributeError:
-        raise requests.exceptions.InvalidURL(f'Invalid URL provided: {url}')
-
     session = requests.Session()
+    session.auth = auth
+
     retry_strategy = Retry(
         total=retries,
         backoff_factor=backoff_factor,
         status_forcelist=[429, 500, 502, 503, 504],
     )
-
     session.mount('https://', HTTPAdapter(max_retries=retry_strategy))
     session.mount('http://', HTTPAdapter(max_retries=retry_strategy))
 
     with session.get(url, stream=True) as s:
+        download_path = _get_download_path(s.url, s.headers.get('content-disposition'), directory)
         s.raise_for_status()
         with open(download_path, "wb") as f:
             for chunk in s.iter_content(chunk_size=chunk_size):